# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/todos": {
      "get": {
        "operationId": "TodoController_findAll",
        "summary": "Todoの一覧が返ってきます",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Todoの一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ]
      }
    },
    "/todos/create": {
      "post": {
        "operationId": "TodoController_create",
        "summary": "Todoを追加します",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todoが追加されました",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoDto"
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ]
      }
    },
    "/todos/update": {
      "put": {
        "operationId": "TodoController_update",
        "summary": "Todoを更新します",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todoが更新されました",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoDto"
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ]
      }
    },
    "/todos/delete": {
      "post": {
        "operationId": "TodoController_delete",
        "summary": "Todoを削除します",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "number"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todoが削除されました",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoDto"
                }
              }
            }
          }
        },
        "tags": [
          "todos"
        ]
      }
    }
  },
  "info": {
    "title": "Todos API",
    "description": "The Todos API description",
    "version": "0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "TodoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "todoのID"
          },
          "title": {
            "type": "string",
            "example": "NestJSの学習",
            "description": "タイトル"
          },
          "completed": {
            "type": "boolean",
            "example": false,
            "description": "todoの完了状態"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-10-05T14:48:00.000Z",
            "description": "作成日"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-10-05T14:48:00.000Z",
            "description": "更新日"
          }
        },
        "required": [
          "id",
          "title",
          "completed",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "NestJSの学習",
            "description": "追加するTodoのタイトル"
          }
        },
        "required": [
          "title"
        ]
      }
    }
  }
}